TP41-218 Move Dcp constants to dcp extension
