+++ edc-extensions/dcp/fx-dcp/src/main/java/org/factoryx/edc/iam/iatp/scope/DefaultScopeExtractor.java
