+++ edc-extensions/dcp/fx-dcp/src/test/java/org/factoryx/edc/iam/iatp/scope/DefaultScopeExtractorTest.java
