+++ settings.gradle.kts
@@ -29,7 +29,7 @@ include(":spi:core-spi")
 // extensions - control plane
 include(":edc-extensions:fx-policy")
 include(":edc-extensions:data-flow-properties-provider")
-include(":edc-extensions:dcp:tx-dcp")
+include(":edc-extensions:dcp:fx-dcp")
 include(":edc-extensions:did-validation-core")
 include(":edc-extensions:contract-validation")
